refactor: 식당 정보 컴포넌트 폴더 이동
